# backend/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WebTui Chat Backend API
  version: 0.9.0
  description: Contract nền cho backend WebTui Chat. Tài liệu API được cập nhật theo từng phase backend.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/oidc/callback:
    get:
      security: []
      summary: Verify the IdP callback and redirect with a one-time completion code
      parameters:
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
      responses:
        "303":
          description: Same-origin redirect to return_to with oidc_code or oidc_error
...
  /api/v1/workspaces/{workspace_id}/api-tokens:
    get:
      summary: Danh sách API token trong workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Danh sách API token
    post:
      summary: Tạo API token mới
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPITokenRequest"
      responses:
        "201":
          description: API token được tạo. Secret chỉ trả về một lần.
  /api/v1/workspaces/{workspace_id}/api-tokens/{token_id}:
    delete:
      summary: Thu hồi API token
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/APITokenID"
      responses:
        "204":
          description: Đã thu hồi API token
...
  /api/v1/workspaces/{workspace_id}/audit-logs:
    get:
      summary: Danh sách audit log
      description: Cần quyền `audit.view`.
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - name: actor_user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: entity_type
          in: query
          schema:
            type: string
        - name: entity_id
          in: query
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit log đã lọc
  /api/v1/workspaces/{workspace_id}/backup-jobs:
    get:
      summary: Danh sách backup job
      description: Cần quyền `backup.manage`.
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Danh sách backup job
    post:
      summary: Tạo backup job database
      description: MVP hiện hỗ trợ `backup_type=database`.
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBackupJobRequest"
      responses:
        "201":
          description: Backup job đã được tạo
  /api/v1/workspaces/{workspace_id}/backup-jobs/{backup_job_id}/runs:
    get:
      summary: Lịch sử chạy backup job
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/BackupJobID"
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Danh sách backup run
  /api/v1/workspaces/{workspace_id}/backup-jobs/{backup_job_id}/run:
    post:
      summary: Chạy backup job thủ công
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/BackupJobID"
      responses:
        "201":
          description: Backup run đã được tạo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    WorkspaceID:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    APITokenID:
      name: token_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    BackupJobID:
      name: backup_job_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    CreateAPITokenRequest:
      type: object
      required: [name, scopes]
      properties:
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
          example: ["message.write", "webhook.write"]
        expires_days:
          type: integer
          description: Số ngày hết hạn. Không gửi hoặc bằng 0 nghĩa là không đặt hạn.
...
    CreateBackupJobRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        target:
          type: string
          enum: [local, minio, s3]
        backup_type:
          type: string
          enum: [database]
        schedule:
          type: string
          description: Tùy chọn. Hỗ trợ cùng định dạng với cronjob.
        status:
          type: string
          enum: [active, paused, disabled]
        config:
          type: object
          additionalProperties: true
...